Move more stuff to dxf2 deprecated package
