+++ dhis-2/dhis-services/dhis-service-dxf2/src/main/java/org/hisp/dhis/dxf2/deprecated/tracker/trackedentity/TrackedEntityInstanceService.java
@@ -39,7 +39,6 @@ import org.hisp.dhis.dxf2.importsummary.ImportSummary;
 import org.hisp.dhis.program.Program;
 import org.hisp.dhis.scheduling.JobConfiguration;
 import org.hisp.dhis.trackedentity.TrackedEntity;
-import org.hisp.dhis.trackedentity.TrackedEntityOuInfo;
 import org.hisp.dhis.trackedentity.TrackedEntityQueryParams;
 import org.hisp.dhis.user.User;
 
@@ -97,7 +96,7 @@ public interface TrackedEntityInstanceService
     TrackedEntityInstance getTrackedEntityInstance( TrackedEntity entityInstance,
         TrackedEntityInstanceParams params, User user );
 
-    List<TrackedEntityOuInfo> getTrackedEntityOuInfoByUid( List<String> uids, User user );
+    List<TrackedEntityOuInfo> getTrackedEntityOuInfoByUid( List<String> uids );
 
     List<TrackedEntityProgramOwnerIds> getTrackedEntityProgramOwnersUidsUsingId( List<Long> teiIds, Program program );
