+++ dhis-2/dhis-services/dhis-service-dxf2/src/main/java/org/hisp/dhis/dxf2/deprecated/tracker/trackedentity/store/TrackedEntityInstanceStore.java
@@ -35,9 +35,8 @@ import org.hisp.dhis.dxf2.deprecated.tracker.trackedentity.Attribute;
 import org.hisp.dhis.dxf2.deprecated.tracker.trackedentity.ProgramOwner;
 import org.hisp.dhis.dxf2.deprecated.tracker.trackedentity.Relationship;
 import org.hisp.dhis.dxf2.deprecated.tracker.trackedentity.TrackedEntityInstance;
+import org.hisp.dhis.dxf2.deprecated.tracker.trackedentity.TrackedEntityOuInfo;
 import org.hisp.dhis.dxf2.deprecated.tracker.trackedentity.TrackedEntityProgramOwnerIds;
-import org.hisp.dhis.trackedentity.TrackedEntityOuInfo;
-import org.hisp.dhis.user.User;
 
 import com.google.common.collect.Multimap;
 
@@ -96,7 +95,7 @@ public interface TrackedEntityInstanceStore
      */
     Multimap<String, String> getOwnedTeis( List<Long> ids, AggregateContext ctx );
 
-    List<TrackedEntityOuInfo> getTrackedEntityOuInfoByUid( List<String> uids, User user );
+    List<TrackedEntityOuInfo> getTrackedEntityOuInfoByUid( List<String> uids );
 
     List<TrackedEntityProgramOwnerIds> getTrackedEntityProgramOwnersUids( List<Long> teiIds, long programId );
 }
